refactor: move browser bindings into own package
