+++ packages/bindings-browser/tsconfig.json
@@ -1,0 +1,18 @@
+{
+	"extends": "../../tsconfig.base.json",
+	"compilerOptions": {
+		"rootDir": "src",
+		"outDir": "build",
+		"module": "ESNext",
+		"moduleResolution": "bundler",
+		"target": "ESNext",
+		"lib": ["DOM", "ESNext"],
+		"customConditions": ["browser"]
+	},
+	"references": [
+		{
+			"path": "../shared/tsconfig.build.json"
+		}
+	],
+	"include": ["src/**/*.ts"]
+}
