+++ packages/core/tsconfig.json
@@ -2,6 +2,8 @@
 {
 	"extends": "../../tsconfig.base.json",
 	"compilerOptions": {},
-	"include": ["src/**/*.ts"],
+	"include": [
+		"src/**/*.ts"
+	],
 	"exclude": ["build/**", "node_modules/**"]
 }
