+++ packages/web/src/script.ts
@@ -1,292 +1,8 @@
+import { db } from "@zwave-js/bindings-browser/db";
+import { fs } from "@zwave-js/bindings-browser/fs";
 import { log as createLogContainer } from "@zwave-js/core/bindings/log/browser";
 import { Bytes } from "@zwave-js/shared";
-import {
-	type Database,
-	type DatabaseFactory,
-	type FSStats,
-	type FileHandle,
-	type FileSystem,
-} from "@zwave-js/shared/bindings";
 import { Driver } from "zwave-js";
-
-const OBJECT_STORE_FILES = "files";
-const OBJECT_STORE_CACHE = "cache";
-
-// IndexedDB-Datenbank öffnen oder erstellen
-function openDatabase(): Promise<IDBDatabase> {
-	return new Promise((resolve, reject) => {
-		const request = indexedDB.open("filesystem", 1);
-
-		request.onupgradeneeded = (event) => {
-			const db = (event.target as IDBOpenDBRequest).result;
-			if (!db.objectStoreNames.contains(OBJECT_STORE_FILES)) {
-				db.createObjectStore(OBJECT_STORE_FILES, { keyPath: "path" });
-			}
-			if (!db.objectStoreNames.contains(OBJECT_STORE_CACHE)) {
-				db.createObjectStore(OBJECT_STORE_CACHE, {
-					keyPath: ["filename", "valueid"],
-				});
-			}
-		};
-
-		request.onsuccess = (event) => {
-			resolve((event.target as IDBOpenDBRequest).result);
-		};
-
-		request.onerror = (event) => {
-			reject((event.target as IDBOpenDBRequest).error);
-		};
-	});
-}
-
-const db = await openDatabase();
-
-// Datei erstellen oder schreiben
-async function writeFile(path: string, data: Uint8Array): Promise<void> {
-	const transaction = db.transaction(OBJECT_STORE_FILES, "readwrite");
-	const store = transaction.objectStore(OBJECT_STORE_FILES);
-
-	const request = store.put({ path, data });
-
-	return new Promise((resolve, reject) => {
-		request.onsuccess = () => resolve();
-		request.onerror = () => reject(request.error);
-	});
-}
-
-// Datei lesen
-async function readFile(path: string): Promise<Uint8Array> {
-	const transaction = db.transaction(OBJECT_STORE_FILES, "readonly");
-	const store = transaction.objectStore(OBJECT_STORE_FILES);
-
-	const request = store.get(path);
-
-	return new Promise((resolve, reject) => {
-		request.onsuccess = () => {
-			const result = request.result;
-			if (result) {
-				resolve(result.data);
-			} else {
-				reject(new Error(`File ${path} not found`));
-			}
-		};
-		request.onerror = () => reject(request.error);
-	});
-}
-
-async function listKeysWithPrefix(prefix: string): Promise<string[]> {
-	const transaction = db.transaction(OBJECT_STORE_FILES, "readonly");
-	const store = transaction.objectStore(OBJECT_STORE_FILES);
-
-	return new Promise((resolve, reject) => {
-		const keys: string[] = [];
-		const request = store.openCursor();
-
-		request.onsuccess = (event) => {
-			const cursor =
-				(event.target as IDBRequest<IDBCursorWithValue>).result;
-			if (cursor) {
-				if (
-					typeof cursor.key === "string"
-					&& cursor.key.startsWith(prefix)
-				) {
-					keys.push(cursor.key as string);
-				}
-				cursor.continue();
-			} else {
-				resolve(keys);
-			}
-		};
-
-		request.onerror = () => reject(request.error);
-	});
-}
-
-async function deleteKeysWithPrefix(prefix: string): Promise<void> {
-	const transaction = db.transaction(OBJECT_STORE_FILES, "readwrite");
-	const store = transaction.objectStore(OBJECT_STORE_FILES);
-
-	return new Promise((resolve, reject) => {
-		const request = store.openCursor();
-
-		request.onsuccess = (event) => {
-			const cursor =
-				(event.target as IDBRequest<IDBCursorWithValue>).result;
-			if (cursor) {
-				if (
-					typeof cursor.key === "string"
-					&& cursor.key.startsWith(prefix)
-				) {
-					cursor.delete();
-				}
-				cursor.continue();
-			} else {
-				resolve();
-			}
-		};
-
-		request.onerror = () => reject(request.error);
-	});
-}
-
-const webFS: FileSystem = {
-	readFile(path) {
-		return readFile(path);
-	},
-	writeFile(path, data) {
-		return writeFile(path, data);
-	},
-	async copyFile(source, dest) {
-		const data = await readFile(source);
-		await writeFile(dest, data);
-	},
-	open: function(
-		path: string,
-		flags: {
-			read: boolean;
-			write: boolean;
-			create: boolean;
-			truncate: boolean;
-		},
-	): Promise<FileHandle> {
-		throw new Error("Function not implemented.");
-	},
-	readDir(path) {
-		return listKeysWithPrefix(path);
-	},
-	stat: function(path: string): Promise<FSStats> {
-		throw new Error("Function not implemented.");
-	},
-	ensureDir: function(path: string): Promise<void> {
-		// No need to create directories
-	},
-	deleteDir(path) {
-		return deleteKeysWithPrefix(path);
-	},
-	makeTempDir: function(prefix: string): Promise<string> {
-		throw new Error("Function not implemented.");
-	},
-};
-
-class IndexedDBBackedCache<V> implements Database<V> {
-	private filename: string;
-	private cache: Map<string, { value: V; timestamp?: number }> = new Map();
-
-	constructor(filename: string) {
-		this.filename = filename;
-	}
-
-	async open(): Promise<void> {
-		const transaction = db.transaction(OBJECT_STORE_CACHE, "readonly");
-		const store = transaction.objectStore(OBJECT_STORE_CACHE);
-		const request = store.openCursor();
-
-		return new Promise((resolve, reject) => {
-			request.onsuccess = (event) => {
-				const cursor =
-					(event.target as IDBRequest<IDBCursorWithValue>).result;
-				if (cursor) {
-					if (cursor.value.filename === this.filename) {
-						this.cache.set(cursor.value.valueid, {
-							value: cursor.value.value,
-							timestamp: cursor.value.timestamp,
-						});
-					}
-					cursor.continue();
-				} else {
-					resolve();
-				}
-			};
-			request.onerror = () => reject(request.error);
-		});
-	}
-
-	close(): Promise<void> {
-		this.cache.clear();
-		return Promise.resolve();
-	}
-
-	has(key: string): boolean {
-		return this.cache.has(key);
-	}
-
-	get(key: string): V | undefined {
-		return this.cache.get(key)?.value;
-	}
-
-	set(key: string, value: V, updateTimestamp: boolean = true): this {
-		const entry = {
-			value,
-			timestamp: updateTimestamp
-				? Date.now()
-				: this.cache.get(key)?.timestamp,
-		};
-		this.cache.set(key, entry);
-
-		// Update IndexedDB in the background
-		const transaction = db.transaction(OBJECT_STORE_CACHE, "readwrite");
-		const store = transaction.objectStore(OBJECT_STORE_CACHE);
-		store.put({
-			filename: this.filename,
-			valueid: key,
-			value,
-			timestamp: entry.timestamp,
-		});
-
-		return this;
-	}
-
-	delete(key: string): boolean {
-		const result = this.cache.delete(key);
-
-		// Update IndexedDB in the background
-		const transaction = db.transaction(OBJECT_STORE_CACHE, "readwrite");
-		const store = transaction.objectStore(OBJECT_STORE_CACHE);
-		store.delete([this.filename, key]);
-
-		return result;
-	}
-
-	clear(): void {
-		this.cache.clear();
-
-		// Update IndexedDB in the background
-		const transaction = db.transaction(OBJECT_STORE_CACHE, "readwrite");
-		const store = transaction.objectStore(OBJECT_STORE_CACHE);
-		const request = store.openCursor();
-
-		request.onsuccess = (event) => {
-			const cursor =
-				(event.target as IDBRequest<IDBCursorWithValue>).result;
-			if (cursor) {
-				if (cursor.value.filename === this.filename) {
-					cursor.delete();
-				}
-				cursor.continue();
-			}
-		};
-	}
-
-	getTimestamp(key: string): number | undefined {
-		return this.cache.get(key)?.timestamp;
-	}
-
-	get size(): number {
-		return this.cache.size;
-	}
-
-	keys() {
-		return this.cache.keys();
-	}
-
-	*entries() {
-		return function*() {
-			for (const [key, { value }] of this.cache.entries()) {
-				yield [key, value];
-			}
-		};
-	}
-}
 
 async function init() {
 	let port: SerialPort;
@@ -336,16 +52,10 @@ async function init() {
 
 	const serialBinding = () => Promise.resolve({ source, sink });
 
-	const dbFactory: DatabaseFactory = {
-		createInstance(filename, options) {
-			return new IndexedDBBackedCache(filename);
-		},
-	};
-
 	const d = new Driver(serialBinding, {
 		host: {
-			fs: webFS,
-			db: dbFactory,
+			fs,
+			db,
 			log: createLogContainer,
 			serial: {
 				// no listing, no creating by path!
